+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Spotify Changelog
 
  * Fix artist and track title parsing.
  * Fix track progress bar integration. Issue: tiliado/nuvolaplayer#761
+ * Fix integration of volume bar.
 
 3.6 - July 31st, 2021
 ---------------------